[CSL-2264] Change Toil to use concrete pure monads
According to our newest guidelines, pure code should be written in concrete
monads. This commit changes function in Toil which were using abstract monad
to use concrete monads.

Code in Toil was written in abstract monads for a good reason: it wanted to do
some IO in some cases and be pure in other cases. These IO actions are:
1. Utxo lookup. It was made pure by introducing 'buildUtxo' function to lookup
all inputs in DB in advance and then pass built utxo to a pure function.
2. Stakes lookup. It was made pure by using free monads.

Type classes like 'MonadUtxo[,Read]', 'MonadStakes[,Read]' and similar have
been removed in favor of concrete monads.

These changes appeared to be necessary for exception handling refactoring,
because I had to drop 'MonadError' constraints and in order to do it I had to
use concrete monads.